tts-sidecar/requirements.txt
soprano-tts
uvicorn
requests
